#93 Refactor: DeliveryValidator 별도의 클래스로 추출
